fix: types and condition statement
